+++ lib/storybook_toolkit_tester.dart
@@ -1,7 +1,5 @@
 library storybook_toolkit_tester;
 
-export 'devices.dart';
-
 import 'dart:async';
 
 import 'package:adaptive_golden_test/adaptive_golden_test.dart';
@@ -9,31 +7,22 @@ import 'package:flutter/material.dart';
 import 'package:recase/recase.dart';
 import 'package:meta/meta.dart';
 import 'package:storybook_toolkit/storybook_toolkit.dart';
-import 'package:storybook_toolkit_tester/devices.dart';
 
 @isTest
 Future<void> testStorybook(
   Storybook storybook, {
-  Set<Device> devices = const {Device.pixel5, Device.iPhone8, Device.iPhone13},
+  Set<DeviceInfo>? devices,
+  bool isFrameVisible = false,
   bool Function(Story story)? filterStories,
   String Function(PathContext)? goldenPathBuilder,
   String rootPath = 'goldens',
 }) async {
-  AdaptiveTestConfiguration.instance.setDeviceVariants({
-    for (Device device in devices)
-      WindowConfigData(
-        device.name,
-        size: device.size,
-        pixelDensity: device.pixelDensity,
-        targetPlatform: device.targetPlatform,
-        borderRadius: device.borderRadius,
-        safeAreaPadding: device.safeAreaPadding,
-        keyboardSize: device.keyboardSize,
-        homeIndicator: device.homeIndicator,
-        notchSize: device.notchSize,
-        punchHole: device.punchHole,
-      ),
-  });
+  AdaptiveTestConfiguration.instance.setDeviceVariants(devices ??
+      {
+        Devices.ios.iPhoneSE,
+        Devices.ios.iPhone12,
+        Devices.android.samsungGalaxyS20,
+      });
   await loadFonts();
   setupFileComparatorWithThreshold();
 
@@ -43,28 +32,25 @@ Future<void> testStorybook(
       (tester, variant) async {
         await tester.pumpWidget(
           AdaptiveWrapper(
-            windowConfig: variant,
-            tester: tester,
-            child: Directionality(
-              textDirection: TextDirection.ltr,
-              child: Storybook(
-                initialStory: story.name,
-                showPanel: false,
-                wrapperBuilder: storybook.wrapperBuilder,
-                stories: storybook.stories,
-              ),
+            device: variant,
+            orientation: Orientation.portrait,
+            isFrameVisible: isFrameVisible,
+            showVirtualKeyboard: false,
+            child: Storybook(
+              initialStory: story.name,
+              showPanel: false,
+              wrapperBuilder: storybook.wrapperBuilder,
+              stories: storybook.stories,
             ),
           ),
         );
 
-        //await tester.expectGolden(variant);
-        //await tester.tap(find.byKey(ValueKey("TestField")));
         if (story.loadDuration != null) await tester.pumpAndSettle(story.loadDuration!);
         await tester.expectGolden(
           variant,
           pathBuilder: (_) {
             final path = story.name.split('/').map((e) => e.snakeCase).join('/');
-            final fileName = "${variant.name}.png";
+            final fileName = "${variant.name.replaceAll(' ', '_')}.png";
 
             final PathContext context = (rootPath: rootPath, path: path, file: fileName);
             return story.goldenPathBuilder?.call(context) ??